Modified the onGoingclasses function to show the next class if no class is ongoing. (#5)
* Update README.md

Added instructions to install to windows.

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Updated the classesOngoing function

Now it returns the immediate next class if no class is ongoing. As a result it also notifies if all classes for the day are over.

* Update README.md

Documentation Improved :D

* Updated the classesOngoing function

Now it returns the immediate next class if no class is ongoing. As a result it also notifies if all classes for the day are over.

* Update README.md

Streamlined the readme to match the recent update.

* Fixed the readme

* Made this perfectly Functioning

Ready for merge